Add parameter for change in enrichment factor.
Generally, the positive test fraction seems to vary inversely with the
number of tests performed, suggesting that the more testing we do, the
more uninfected people we sweep up in the test program.  This additional
parameter represents this dependence in a simple model where an e-fold
increase in the effective number of tests produces a linear decrease in
the enrichment factor.